+++ src/main/java/com/demo/challenge/dtos/ProductDTO.java
@@ -2,7 +2,7 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package com.demo.challenge.dto;
+package com.demo.challenge.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
