+++ src/main/java/com/demo/challenge/dtos/ProductSaleDTO.java
@@ -1,4 +1,4 @@
-package com.demo.challenge.dto;
+package com.demo.challenge.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
