+++ src/main/java/com/demo/challenge/dtos/SaleDTO.java
@@ -1,6 +1,5 @@
-package com.demo.challenge.dto;
+package com.demo.challenge.dtos;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -8,7 +7,6 @@ import lombok.Setter;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.ArrayList;
 
 import java.util.List;
