+++ src/main/java/com/demo/challenge/repositories/ICustomerRepository.java
@@ -2,9 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package com.demo.challenge.repository;
+package com.demo.challenge.repositories;
 
-import com.demo.challenge.entitys.Customer;
+import com.demo.challenge.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
