+++ src/main/java/com/demo/challenge/repositories/IProductRepository.java
@@ -2,11 +2,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package com.demo.challenge.repository;
+package com.demo.challenge.repositories;
 
-import com.demo.challenge.dto.ProductDTO;
-import com.demo.challenge.dto.ProductProviderDTO;
-import com.demo.challenge.entitys.Product;
+import com.demo.challenge.entities.Product;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
