+++ src/main/java/com/demo/challenge/repositories/IProviderRepository.java
@@ -2,13 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package com.demo.challenge.repository;
+package com.demo.challenge.repositories;
 
-import com.demo.challenge.entitys.Provider;
+import com.demo.challenge.entities.Provider;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 /**
  *
