+++ src/main/java/com/demo/challenge/repositories/ISaleRepository.java
@@ -2,11 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
  */
-package com.demo.challenge.repository;
+package com.demo.challenge.repositories;
 
-import com.demo.challenge.entitys.Sale;
+import com.demo.challenge.entities.Sale;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
