+++ src/main/java/com/demo/challenge/servicesInterfaces/IProviderService.java
@@ -4,8 +4,8 @@
  */
 package com.demo.challenge.servicesInterfaces;
 
-import com.demo.challenge.dto.ProviderDTO;
-import com.demo.challenge.entitys.Provider;
+import com.demo.challenge.dtos.ProviderDTO;
+import com.demo.challenge.entities.Provider;
 import org.springframework.http.ResponseEntity;
 
 import java.util.List;
