+++ src/main/java/com/demo/challenge/servicesInterfaces/ISaleService.java
@@ -4,8 +4,8 @@
  */
 package com.demo.challenge.servicesInterfaces;
 
-import com.demo.challenge.dto.SaleDTO;
-import com.demo.challenge.entitys.Sale;
+import com.demo.challenge.dtos.SaleDTO;
+
 import java.time.LocalDate;
 import java.util.List;
